requirements.txt
# Core data handling
pandas>=2.2.0
numpy>=1.26.0
lightgbm>=4.0.0
joblib>=1.3.0
streamlit>=1.30.0
python-dotenv
openpyxl


# Modeling
scikit-learn>=1.5.0

# Database connection (optional, for SQL Server)
sqlalchemy>=2.0.0
pyodbc>=5.0.0

